refactor(cashu): move wallet + mint-validation + subscription managers into @agicash/cashu
Completes the framework-free @agicash/cashu lib (Chunk C of the cashu lib-ify):

- utils.ts (ExtendedCashuWallet, getCashuWallet, unit/keyset/mint-url helpers),
  mint-validation.ts, and the melt-quote/mint-quote subscription MANAGER classes
  -> packages/cashu/src (byte-identical git mv)
- the managers now import getCashuWallet from concrete ./utils (NOT the ~/lib/cashu barrel,
  which re-exports a React hook) and isSubset from @agicash/utils, so the lib stays React-free
- isSubset relocated from app/lib/utils.ts into @agicash/utils (single source); app/lib/utils.ts
  re-exports it so ~/lib/utils consumers (the proof-state manager) are untouched
- cashu utils money types now from @agicash/utils/money; type-fest added to @agicash/cashu
- old ~/lib/cashu/{utils,mint-validation,*-subscription-manager} kept as transitional shims
- proof-state-subscription-manager.ts (features/send): added a note that it likely belongs in
  @agicash/cashu too, deferred to the send/swap phase (it depends on send-domain types)

Web keeps the React hook (useOnMeltQuoteStateChange) + animated-qr. typecheck, biome, build all
green; 133 tests preserved; FRAMEWORK-FREE GATE verified (zero react/@tanstack/DOM in
@agicash/cashu).
