+++ apps/web-wallet/app/lib/utils.ts
@@ -9,24 +9,9 @@ export function sum(numbers: number[]) {
   return numbers.reduce((acc, curr) => acc + curr, 0);
 }
 
-export function isSubset<T>(subset: Set<T>, superset: Set<T>): boolean {
-  const isSubsetOf = (
-    subset as unknown as {
-      isSubsetOf?: (other: ReadonlySet<T>) => boolean;
-    }
-  ).isSubsetOf;
-
-  if (typeof isSubsetOf === 'function') {
-    return isSubsetOf.call(subset, superset);
-  }
-
-  for (const item of subset) {
-    if (!superset.has(item)) {
-      return false;
-    }
-  }
-  return true;
-}
+// isSubset moved to @agicash/utils (shared with the @agicash/cashu subscription
+// managers); re-exported here so ~/lib/utils consumers stay unchanged.
+export { isSubset } from '@agicash/utils';
 
 export function isObject(v: unknown): v is Record<string, unknown> {
   return v !== null && typeof v === 'object' && !Array.isArray(v);
