+++ packages/cashu/package.json
@@ -14,6 +14,7 @@
     "@agicash/utils": "workspace:*",
     "@cashu/cashu-ts": "catalog:",
     "@noble/hashes": "catalog:",
+    "type-fest": "catalog:",
     "zod": "catalog:"
   },
   "devDependencies": {
