+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
 export * from './money';
 export * from './json';
 export * from './zod';
+export * from './collections';
